requirements.txt
Flask==3.1.0
requests==2.32.3
pyinstaller==6.11.1
pywebview==5.4
